Make CMU-MOSEI loadable on a 16GB machine
Naive loading needed ~8GB for a 4.4GB archive: reading the pickle held it
alongside the converted tensors, and compute_stats and normalize_features each
allocated a second full-size copy. Now arrays are dropped from the pickle as
they convert, statistics accumulate in float64 over chunks, and normalisation
writes back in place.

Adds storage_dtype (float16 for MOSEI) which halves the cache to 2.0GB.
Features are z-scored before storage and widened to float32 per batch, so no
corruption operator or model ever does float16 arithmetic.

Freezes the MOSEI evaluation split: 16326/1871/4659.

+++ src/wfb/data/datamodule.py
@@ -54,7 +54,10 @@ class MultimodalDataset(Dataset[Batch]):
         return len(self.split)
 
     def __getitem__(self, index: int) -> Batch:
-        features = {m: t[index] for m, t in self.split.features.items()}
+        # Widen to float32 here. Large corpora may be cached as float16 to fit in memory,
+        # but corruption operators and models must not do float16 arithmetic — the noise
+        # calibration and the z-scored feature range both assume float32 precision.
+        features = {m: t[index].to(torch.float32) for m, t in self.split.features.items()}
         if not self.plan.is_clean:
             features = apply_plan(
                 features,
